+++ Source/OpenLight/BRDF/Fresnel.h
@@ -7,9 +7,9 @@ class Fresnel
 public:
 	virtual ~Fresnel();
 
-	virtual Spectrum Evalute( double cosi ) const = 0;
+	virtual Spectrum Evalute( float cosi ) const = 0;
 };
 
-Spectrum FresnelDieletricIntoConductor( double CosThetaI , const Spectrum& etaI , const Spectrum& etaT , const Spectrum& k );
+Spectrum FresnelDieletricIntoConductor( float CosThetaI , const Spectrum& etaI , const Spectrum& etaT , const Spectrum& k );
 
-double FresnelDielectricIntoDielectric( double CosThetaI , double etaI , double etaT );
+float FresnelDielectricIntoDielectric( float CosThetaI , float etaI , float etaT );
